+++ lib/src/handler/w3c/window.dart
@@ -65,7 +65,7 @@ class W3cWindowHandler extends WindowHandler {
 
   @override
   Rect parseRectResponse(WebDriverResponse response) {
-    final rect = parseW3cResponse(response);
+    final rect = parseW3cResponse(response) as Map<String, Object?>;
     return Rect(
       left: (rect['x'] as num).toInt(),
       top: (rect['y'] as num).toInt(),
@@ -142,15 +142,21 @@ class W3cWindowHandler extends WindowHandler {
       WebDriverRequest.postRequest('execute/sync', {
         'script':
             'return { width: window.innerWidth, height: window.innerHeight };',
-        'args': []
+        'args': <Object>[]
       });
 
   @override
+<<<<<<< HEAD
   Size parseInnerSizeResponse(WebDriverResponse response) {
     final size = parseW3cResponse(response);
     return Size(
       width: (size['width'] as num).toInt(),
       height: (size['height'] as num).toInt(),
     );
+=======
+  Rectangle<int> parseInnerSizeResponse(WebDriverResponse response) {
+    final size = parseW3cResponse(response) as Map<String, Object?>;
+    return Rectangle(0, 0, size['width'] as int, size['height'] as int);
+>>>>>>> master
   }
 }