+++ lib/src/sync/web_driver.dart
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 import 'dart:convert' show base64;
+import 'dart:typed_data';
 
 import '../../async_core.dart' as async_core;
 import '../common/by.dart';
@@ -117,7 +118,9 @@ class WebDriver implements SearchContext {
   }
 
   /// Search for an element within the entire current page.
-  /// Throws [NoSuchElementException] if a matching element is not found.
+  ///
+  /// Throws [async_core.NoSuchElementException] if a matching element is not
+  /// found.
   @override
   WebElement findElement(By by) => WebElement(
       this,
@@ -129,7 +132,8 @@ class WebDriver implements SearchContext {
       by);
 
   /// Search for an element by xpath within the entire current page.
-  /// Throws [NoSuchElementException] if a matching element is not found.
+  /// Throws [async_core.NoSuchElementException] if a matching element is not
+  /// found.
   WebElement findElementByXpath(String by) => findElement(By.xpath(by));
 
   /// An artist's rendition of the current page's source.
@@ -213,13 +217,13 @@ class WebDriver implements SearchContext {
       _handler.core.parseScreenshotResponse);
 
   /// Take a screenshot of the current page as PNG as list of uint8.
-  List<int> captureScreenshotAsList() {
+  Uint8List captureScreenshotAsList() {
     final base64Encoded = captureScreenshotAsBase64();
     return base64.decode(base64Encoded);
   }
 
   /// Take a screenshot of the specified element as PNG as list of uint8.
-  List<int> captureElementScreenshotAsList(WebElement element) {
+  Uint8List captureElementScreenshotAsList(WebElement element) {
     final base64Encoded = captureElementScreenshotAsBase64(element);
     return base64.decode(base64Encoded);
   }