+++ test/async_window_test.dart
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @TestOn('vm')
-library webdriver.window_test;
+library;
 
 import 'dart:async';
 
@@ -36,7 +36,7 @@ void main() {
       const size = Size(width: 600, height: 400);
 
       // Firefox may take a bit longer to do the resize.
-      await Future.delayed(const Duration(seconds: 1));
+      await Future<void>.delayed(const Duration(seconds: 1));
       await window.setSize(size);
       expect(await window.size, size);
     });