+++ test/sync/logs.dart
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @TestOn('vm')
-library webdriver.logs_test;
+library;
 
 import 'package:test/test.dart';
 import 'package:webdriver/sync_core.dart';